// src/Moryx.Operators.Web/app/src/app/api/models/extended-operator-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AttendableResourceModel } from '../models/attendable-resource-model';
export interface ExtendedOperatorModel {
  assignedResources?: Array<AttendableResourceModel> | null;
  firstName?: string | null;
  identifier?: string | null;
  lastName?: string | null;
  pseudonym?: string | null;
}